+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.56.3](https://github.com/elephant-xyz/elephant-cli/compare/v1.56.2...v1.56.3) (2025-12-03)
+
+
+### Bug Fixes
+
+* shouldn't create property exemption and jurisdiction relation ([#228](https://github.com/elephant-xyz/elephant-cli/issues/228)) ([af0d040](https://github.com/elephant-xyz/elephant-cli/commit/af0d0400c86df4bb6f3da2b49e49e1685cabd5a7))
+
+## [1.56.2](https://github.com/elephant-xyz/elephant-cli/compare/v1.56.1...v1.56.2) (2025-12-02)
+
+
+### Bug Fixes
+
+* add tax jurisdiction / exemption relationship with factsheet ([#227](https://github.com/elephant-xyz/elephant-cli/issues/227)) ([6422878](https://github.com/elephant-xyz/elephant-cli/commit/6422878d4ca1b0bd6a70a8c76aa4e1390f79d80f))
+
 ## [1.56.1](https://github.com/elephant-xyz/elephant-cli/compare/v1.56.0...v1.56.1) (2025-12-02)